refactor(p1): extract SystemAccessShared.kt — trackSystemTool + schema DSL + masking primitives
First step of splitting up the 1441-line SystemAccessTools god class. Cross-domain
helpers (used by 3+ tool families) lift out to a new SystemAccessShared.kt sibling;
single-caller helpers stay put for now and will travel with their tool when the
per-domain factories land in follow-up commits.

What moves to SystemAccessShared.kt (141 lines new):
- internal class SystemAccessDeps(activityStore, permissionBroker) — 2-field bundle
  so per-domain factories only thread one parameter instead of two
- internal suspend fun SystemAccessDeps.trackSystemTool(...) — ensureGranted →
  startTool → block → complete/fail, behavior-identical to the prior class method
- internal fun obj / stringProp / booleanProp / integerProp / enumProp — schema DSL
- internal fun JsonElement.limit / .safePreview — input parsers
- internal fun maskPhone / maskEmail — PII masking used by 6+ tools
- private fun List<UIMessagePart>.previewText — sole caller is trackSystemTool

What stays in SystemAccessTools.kt (1441 → 1338 lines):
- JsonElement.timeMillis (only calendar_create uses it)
- appLabel (only usage_stats_list uses it)
- mimeFromExtension (only share_file uses it)
- redactMiddle (only wifi_status uses it)
- All domain query/create helpers (queryContacts, querySms, etc.)
- All 27 tool definitions

Mechanical churn: added `private val deps = SystemAccessDeps(...)` field on the
class; rewrote 28 call sites from `trackSystemTool(...)` to `deps.trackSystemTool(...)`;
dropped now-dead imports (InputSchema, kotlin.math.min, kotlinx.serialization.json.jsonObject,
me.rerere.ai.ui.UIMessagePart).

NOTE on package-wide duplication: this same schema DSL is independently re-declared
as `private` members in 9 other tool files in the agent/tools/ package. Consolidating
those into ToolJson.kt is a separate cleanup; intentionally out of scope here so the
god-class split keeps blast radius minimal. The Shared.kt section comment flags this.

Verification:
- :app:compileDebugKotlin --offline → BUILD SUCCESSFUL
- :app:testDebugUnitTest --offline → 427 tests, 6 failed (baseline preserved:
  ContextFootprintEstimatorTest×3, GenerativeUiPlannerTest×2, DefaultProvidersTest×1)

Sub-agent review: APPROVE WITH NITS (misleading file comment — fixed before commit).
